+++ eflo-controller-20221215/alibabacloud_eflo_controller20221215/models/_list_tag_resources_response_body.py
@@ -18,7 +18,6 @@ class ListTagResourcesResponseBody(DaraModel):
         self.next_token = next_token
         # The request ID.
         self.request_id = request_id
-        # The tags.
         self.tag_resources = tag_resources
 
     def validate(self):
@@ -98,18 +97,9 @@ class ListTagResourcesResponseBodyTagResourcesTagResource(DaraModel):
         tag_key: str = None,
         tag_value: str = None,
     ):
-        # The resource ID.
         self.resource_id = resource_id
-        # The resource type.
-        # 
-        # Valid values:
-        # 
-        # *   Node
-        # *   Cluster
         self.resource_type = resource_type
-        # The tag key.
         self.tag_key = tag_key
-        # The tag value.
         self.tag_value = tag_value
 
     def validate(self):